seed-init-service/requirements.txt
Flask
google-cloud-kms
google-cloud-secret-manager
